- Added subscriptions to all 3 event topics in the client - Implemented the streams for all 3 types of events - Added a "powerDown" method to the movement service - Fixed the telemetry actor to gather correct telemetry data (the old version wasn't providing correct data) - Added filtering to the vision actor to filter out too small objects - Fixed the initialization of the SPI interface for the vision actor

+++ server/telemetry/src/main/java/de/codecentric/iot/rapiro/telemetry/model/TelemetryData.java
@@ -6,7 +6,7 @@ package de.codecentric.iot.rapiro.telemetry.model;
 public class TelemetryData {
 
     private double cpuLoad;
-    private long memoryUsage;
+    private double memoryUsage;
 
     public TelemetryData() {
     }
@@ -24,11 +24,11 @@ public class TelemetryData {
         this.cpuLoad = cpuLoad;
     }
 
-    public long getMemoryUsage() {
+    public double getMemoryUsage() {
         return memoryUsage;
     }
 
-    public void setMemoryUsage(long memoryUsage) {
+    public void setMemoryUsage(double memoryUsage) {
         this.memoryUsage = memoryUsage;
     }
 }
